requirements.txt
absl-py==2.3.1
ale-py==0.11.2
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.20.0
fonttools==4.61.0
fsspec==2025.10.0
grpcio==1.76.0
gym==0.26.2
gym-notices==0.1.0
gymnasium==1.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
massive==2.0.2
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
polygon==1.2.8
protobuf==6.33.1
psutil==7.1.3
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rich==14.2.0
sb3_contrib==2.7.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
stable_baselines3==2.7.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.9.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
wheel==0.45.1
